+++ gui/gui.h
@@ -18,7 +18,7 @@ SDL_Renderer *render_text(const char *text, int x, int y,
 
 SDL_Renderer *render_scope(char *text, int x, int y, SDL_Renderer *renderer);
 
-int create_scope(double *signal, int layout, int size);
+int create_scope(float *signal, int layout, int size);
 
 int create_static_plot(int layout, int size, double *signal);
